+++ src/common/structs/includes.ts
@@ -8,8 +8,8 @@ export interface StatefulCounts<K extends string> {
   available: number;
 }
 
-/** An owner is an identity that created a resource. */
-export interface OwnerInclude {
+/** A creator is an identity that created a resource. */
+export interface UserIncludes {
   accounts?: {
     [key: string]: PublicAccount;
   };