+++ src/resources/containers/instances/ssh.ts
@@ -1,9 +1,4 @@
-import {
-  Resource,
-  ResourceId,
-  OwnerScope,
-  Time,
-} from "../../../common/structs";
+import { Resource, ResourceId, UserScope, Time } from "../../../common/structs";
 import {
   links,
   getRequest,
@@ -23,7 +18,7 @@ export interface SSHToken extends Resource {
   instance_id: ResourceId;
   container_id: ResourceId;
   hub_id: ResourceId;
-  owner: OwnerScope;
+  creator: UserScope;
   events: {
     created: Time;
     used: Time;