+++ src/resources/surveys/index.ts
@@ -1,0 +1,4 @@
+import * as Response from "./response";
+import * as Survey from "./survey";
+
+export { Response, Survey };