+++ CHANGELOG.md
@@ -9,7 +9,18 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Survey structures
+- Survey link(s)
+- Initialize task action for environments
+
 ### Changed
+
+### Breaking Changes
+
+- Field owner renamed to creator
+- Field owners renamed to creators
+- Interface OwnerScope renamed to CreatorScope
+- Interface OwnerIncludes interface renamed to CreatorIncludes
 
 ## [1.8.6]